Refactor marker icon
